mass deletions
