--- packages/mobile/src/screens/search-results-screen/SearchFocusContext.tsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-
-export const SearchFocusContext = createContext({
-  isFocused: true
-})
