--- packages/mobile/src/screens/search-results-screen/SearchQueryContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react'
-
-export const SearchQueryContext = createContext({
-  query: '',
-  isTagSearch: false
-})
