--- packages/mobile/src/screens/search-results-screen/constants.ts
@@ -1,1 +1,0 @@
-export const ALL_CATEGORY_RESULTS_LIMIT = 15
