--- packages/mobile/src/screens/search-results-screen/index.ts
@@ -1,2 +1,0 @@
-export { SearchResultsScreen } from './SearchResultsScreen'
-export { TagSearchScreen } from './TagSearchScreen'
