--- packages/mobile/src/screens/search-screen-v2/index.ts
@@ -1,2 +1,0 @@
-export { SearchScreenStack } from './SearchScreenV2'
-export type { SearchParams } from './searchParams'
