+++ packages/mobile/src/screens/search-screen/SearchCatalogTile.tsx
