+++ packages/mobile/src/screens/search-screen/SearchCategoriesAndFilters.tsx
