--- packages/mobile/src/screens/search-screen/SearchHistory/index.ts
@@ -1,1 +1,0 @@
-export * from './SearchHistory'
