+++ packages/mobile/src/screens/search-screen/SearchItem.tsx
