--- packages/mobile/src/screens/search-screen/SearchResults/index.ts
@@ -1,1 +1,0 @@
-export * from './SearchResults'
