+++ packages/mobile/src/screens/search-screen/search-results/AllResults.tsx
