+++ packages/mobile/src/screens/search-screen/search-results/SearchResults.tsx
