+++ packages/mobile/src/screens/search-screen/searchParams.ts
