+++ packages/mobile/src/screens/search-screen/searchState.ts
