+++ packages/web/src/components/search/SearchBar.jsx
@@ -5,7 +5,10 @@ import {
   IconArrowRight as IconArrow,
   IconSearch,
   setupHotkeys,
-  removeHotkeys
+  removeHotkeys,
+  IconCloseAlt,
+  IconButton,
+  spacing
 } from '@audius/harmony'
 import AutoComplete from 'antd/lib/auto-complete'
 import Input from 'antd/lib/input'
@@ -95,15 +98,19 @@ class SearchBar extends Component {
 
   onSearch = (value, action) => {
     const trimmedValue = value.slice(0, maxLength)
+    if (trimmedValue === this.state.value) return
     this.setState({ value: trimmedValue, valueFromParent: false })
 
     // Set the search state but don't actually call search
     this.props.onSearch(trimmedValue, false)
-    // Set a debounce timer for 100ms to actually send the search
+    // Set a debounce timer for 400ms to actually send the search
     this.setState({
       debounce: setTimeout(() => {
-        this.props.onSearch(trimmedValue, true)
-      }, 100)
+        this.props.onSearch(trimmedValue, !this.props.isViewingSearchPage)
+        if (this.props.isViewingSearchPage) {
+          this.props.onSubmit(this.state.value)
+        }
+      }, 400)
     })
   }
 
@@ -118,6 +125,20 @@ class SearchBar extends Component {
         this.props.onSubmit(this.state.value)
       }
     } else if (value !== NO_RESULTS_OPTION) {
+      // Have to do this lookup because ant d autocomplete only accepts string values
+      const results = this.props.dataSource.sections.reduce(
+        (result, current) => [...result, ...current.children],
+        []
+      )
+      const recentSearch = results.find(({ key }) => key === value)
+      if (recentSearch) {
+        this.props.addRecentSearch({
+          searchItem: {
+            kind: recentSearch.kind,
+            id: recentSearch.id
+          }
+        })
+      }
       this.props.goToRoute(value)
       if (this.props.onSelect) this.props.onSelect(value)
     }
@@ -191,7 +212,7 @@ class SearchBar extends Component {
           this.setState({ shouldDismissTagPopup: true })
         if (
           (this.props.isTagSearch && this.state.value.length > 1) ||
-          (!this.props.isTagSearch && this.state.value.length > 0)
+          !this.props.isTagSearch
         ) {
           this.props.onSubmit(this.state.value)
           this.setState({ debounce: null })
@@ -361,20 +382,30 @@ class SearchBar extends Component {
     // If we're searching for a tag,
     // don't open the autocomplete popup, and instead show our
     // own custom component below.
-    const showAutocomplete = !isTagSearch && this.state.open
+    const showAutocomplete =
+      !isTagSearch && this.state.open && !this.props.isViewingSearchPage
     const showTagPopup =
       isTagSearch && this.state.open && !this.state.shouldDismissTagPopup
-    return (
-      <div
-        className={styles.searchBar}
-        id='search-bar-autocomplete'
-        ref={this.searchBarRef}
-      >
-        {/* show search spinner if not a tag search and there is some value present */}
-        {!isTagSearch && this.state.value && (
+    const showSpinner = status === Status.LOADING && this.state.open
+
+    const renderSuffix = () => {
+      if (this.state.value && !showSpinner) {
+        return (
+          <IconButton
+            icon={IconCloseAlt}
+            size='2xs'
+            color='subdued'
+            onClick={this.props.onClear}
+          />
+        )
+      }
+
+      /* show search spinner if not a tag search and there is some value present */
+      if (!isTagSearch && this.state.value) {
+        return (
           <div
             className={cn(styles.loadingAnimation, {
-              [styles.show]: status === Status.LOADING && this.state.open
+              [styles.show]: showSpinner
             })}
           >
             <Lottie
@@ -385,7 +416,16 @@ class SearchBar extends Component {
               }}
             />
           </div>
-        )}
+        )
+      }
+    }
+
+    return (
+      <div
+        className={styles.searchBar}
+        id='search-bar-autocomplete'
+        ref={this.searchBarRef}
+      >
         <AutoComplete
           ref={this.autoCompleteRef}
           dropdownClassName={cn(styles.searchBox, {
@@ -408,7 +448,14 @@ class SearchBar extends Component {
             name='search'
             autoComplete='off'
             type='search'
-            prefix={<IconSearch color='subdued' />}
+            css={{ paddingRight: spacing.s }}
+            prefix={
+              <IconSearch
+                color='subdued'
+                onClick={() => this.props.onSubmit('')}
+              />
+            }
+            suffix={renderSuffix()}
             onKeyDown={this.onKeyDown}
             spellCheck={false}
           />
